+++ py/surveysim/simulator.py
@@ -14,6 +14,7 @@ import desiutil.log
 
 import desisurvey.ephemerides
 import desisurvey.afternoonplan
+import desisurvey.schedule
 import desisurvey.plan
 import desisurvey.utils
 import desisurvey.config
@@ -33,34 +34,38 @@ class Simulator(object):
         Survey stops on the morning of this date.
     progress : desisurvey.progress.Progress
         Progress of survey at the start of this simulation.
+    weather : surveysim.weather.Weather
+        Simulated weather conditions use use.
+    stats : astropy.table.Table
+        Table of per-night efficiency statistics to update.
     strategy : str
         Strategy to use for scheduling tiles during each night.
     plan : str or None
         Name of plan file to use. Required unless strategy is 'baseline'.
-    seed : int or None
-        Random number seed used to generate weather conditions.
+    gen : numpy.random.RandomState or None
+        Random number generator to use for reproducible samples. Will be
+        initialized (un-reproducibly) if None.
+    computeHA : bool
+        ?
     """
-    def __init__(self, start_date, stop_date, progress, strategy='baseline',
-                 plan=None, seed=20190823, computeHA=False):
+    def __init__(self, start_date, stop_date, progress, weather, stats,
+                 strategy='baseline', plan=None, gen=None,
+                 computeHA=False):
         self.log = desiutil.log.get_logger()
-
-        config = desisurvey.config.Configuration()
-        nominal_stop_date = config.last_day()
+        self.config = desisurvey.config.Configuration()
 
         # Validate date range.
-        self.num_days = (stop_date - start_date).days
-        if self.num_days <= 0:
+        if start_date >= stop_date:
             raise ValueError('Expected start_date < stop_date.')
+        if (start_date < self.config.first_day() or
+            stop_date > self.config.last_day()):
+            raise ValueError('Cannot simulate beyond nominal survey dates.')
         self.start_date = start_date
         self.stop_date = stop_date
+        self.last_index = (self.stop_date - self.config.first_day()).days
 
         # Tabulate sun and moon ephemerides for each night of the survey.
-        if computeHA and stop_date < nominal_stop_date:
-            self.ephem = desisurvey.ephemerides.Ephemerides(
-                start_date, nominal_stop_date, use_cache=True)
-        else:
-            self.ephem = desisurvey.ephemerides.Ephemerides(
-                start_date, stop_date, use_cache=True)
+        self.ephem = desisurvey.ephemerides.Ephemerides(use_cache=True)
 
         if strategy == 'baseline':
             # Build the survey plan.
@@ -68,41 +73,37 @@ class Simulator(object):
                 self.ephem.start.mjd, self.ephem.stop.mjd, self.ephem,
                 computeHA)
         else:
-            # Load the survey planner.
-            self.sp = desisurvey.plan.Planner()
+            # Load the survey scheduler to use.
+            self.sp = desisurvey.schedule.Scheduler()
         self.strategy = strategy
 
-        # Initialize the random number generator to use for simulating
-        # the weather and adding jitter to exposure times.
-        self.gen = np.random.RandomState(seed)
-
-        # Initialize the survey weather conditions generator.
-        self.weather = surveysim.weather.Weather(
-            start_date, stop_date, gen=self.gen)
+        self.gen = gen
+        self.weather = weather
 
         if plan is not None:
             # Load the plan to use.
-            config = desisurvey.config.Configuration()
-            self.plan = astropy.table.Table.read(config.get_path(plan))
+            self.plan = astropy.table.Table.read(self.config.get_path(plan))
             assert np.all(self.sp.tiles['tileid'] == self.plan['tileid'])
         else:
             self.plan = None
 
-        # Initialize a table for efficiency tracking.
-        self.etrack = astropy.table.Table()
-        self.etrack['available'] = np.zeros(self.num_days)
-        self.etrack['overhead'] = np.zeros(self.num_days)
-        self.etrack['delay'] = np.zeros(self.num_days)
-        self.etrack['dawn'] = np.zeros(self.num_days)
-        self.etrack['live'] = np.zeros(self.num_days)
+        if len(stats) != (self.config.last_day() - self.config.first_day()).days:
+            raise ValueError('Input stats table has wrong length.')
+        self.stats = stats
 
-        self.day_index = 0
+        self.day_index = (self.start_date - self.config.first_day()).days
         self.survey_done = False
         self.completed = progress.completed()
         self.progress = progress
         self.log.info(
             'Will simulate {0} to {1} with {2:.1f} / {3} tiles completed.'
             .format(start_date, stop_date, self.completed, progress.num_tiles))
+
+    @property
+    def date(self):
+        """The current simulation date as a datetime object.
+        """
+        return self.config.first_day() + datetime.timedelta(days=self.day_index)
 
     def next_day(self):
         """Simulate the next day of survey operations.
@@ -116,10 +117,10 @@ class Simulator(object):
         bool
             True if there are more days to simulate.
         """
-        if self.day_index >= self.num_days or self.survey_done:
+        if self.day_index >= self.last_index or self.survey_done:
             return False
 
-        date = self.start_date + datetime.timedelta(days=self.day_index)
+        date = self.date
         self.log.info('Simulating {0}'.format(date))
 
         if desisurvey.utils.is_monsoon(date):
@@ -148,19 +149,26 @@ class Simulator(object):
 
             # Update our efficiency tracker.
             for mode in totals:
-                self.etrack[mode][self.day_index] = totals[mode].to(u.day).value
+                self.stats[mode][self.day_index] = totals[mode].to(u.day).value
 
+            # Progress report.
             completed = self.progress.completed()
-            self.log.info(
-                'Completed {0:.1f} tiles tonight, {1:.1f} remaining.'
-                .format(completed - self.completed,
-                        self.progress.num_tiles - completed))
+            self.log.info('Completed {0:.1f} tiles tonight.'
+                          .format(completed - self.completed))
             self.completed = completed
-            if self.progress.num_tiles - completed < 0.1:
-                self.survey_done = True
+
+            # Are we done yet?
+            if self.plan:
+                # Have we completed a group so that we need to update the plan?
+                self.survey_done = desisurvey.plan.update_required(
+                    self.plan, self.progress)
+            else:
+                # With no plan, keep going until all tiles are observed.
+                if self.progress.num_tiles - completed < 0.1:
+                    self.survey_done = True
 
         self.day_index += 1
-        if self.day_index == self.num_days:
+        if self.day_index == self.last_index:
             self.survey_done = True
 
         return not self.survey_done