+++ py/surveysim/test/test_surveysim.py
@@ -2,40 +2,36 @@ from __future__ import print_function, division, absolute_import
 
 import datetime
 import unittest
-import os
+import tempfile
 import shutil
-import uuid
+import os
 
 import numpy as np
 
-from astropy.table import Table
-from astropy import units
-from astropy.time import Time
-
-import desiutil.log
+import astropy.table
 
 import desisurvey.config
 import desisurvey.progress
 
+import surveysim.weather
+import surveysim.simulator
 
 class TestSurveySim(unittest.TestCase):
 
     @classmethod
     def setUpClass(cls):
-        cls.origdir = os.getcwd()
-        cls.testdir = os.path.abspath('./test-{}'.format(uuid.uuid4()))
-        os.mkdir(cls.testdir)
-        os.chdir(cls.testdir)
-        # Write all outputs to our test path.
-        cls.config = desisurvey.config.Configuration()
-        cls.config.set_output_path(cls.testdir)
+        # Create a temporary directory.
+        cls.tmpdir = tempfile.mkdtemp()
+        # Write output files to this temporary directory.
+        config = desisurvey.config.Configuration()
+        config.set_output_path(cls.tmpdir)
 
     @classmethod
     def tearDownClass(cls):
-        cls.config.reset()
-        os.chdir(cls.origdir)
-        if os.path.exists(cls.testdir):
-            shutil.rmtree(cls.testdir)
+        # Remove the directory after the test.
+        shutil.rmtree(cls.tmpdir)
+        # Reset our configuration.
+        desisurvey.config.Configuration.reset()
 
     def setUp(self):
         pass
@@ -43,18 +39,25 @@ class TestSurveySim(unittest.TestCase):
     def tearDown(self):
         pass
 
-    def test_surveysim(self):
-        from surveysim.simulator import Simulator
+    def test_simulator(self):
+        gen = np.random.RandomState(123)
         start = datetime.date(2019,9,1)
         stop = datetime.date(2019,9,8)
         progress = desisurvey.progress.Progress()
-        sim = Simulator(start, stop, progress, seed=123456)
+        weather = surveysim.weather.Weather(start, stop, gen=gen)
+        # Initialize a table for efficiency stats tracking.
+        stats = astropy.table.Table()
+        config = desisurvey.config.Configuration()
+        num_nights = (config.last_day() - config.first_day()).days
+        stats['available'] = np.zeros(num_nights)
+        stats['overhead'] = np.zeros(num_nights)
+        stats['delay'] = np.zeros(num_nights)
+        stats['dawn'] = np.zeros(num_nights)
+        stats['live'] = np.zeros(num_nights)
+        sim = surveysim.simulator.Simulator(
+            start, stop, progress, weather, stats, gen=gen)
         while sim.next_day():
             pass
-
-        #- A plan should exist for every night
-        for i in range(1,8):
-            self.assertTrue(os.path.exists('obsplan201909{:02d}.fits'.format(i)))
 
         #- Confirm that observations occur only on survey dates.
         obs_table = progress.get_summary('observed')